remove decfp from testing because it doesn't work on 32-bit

+++ test/testfiles.jl
@@ -630,7 +630,7 @@ testfiles = [
         nothing
     ),
     # https://github.com/JuliaData/CSV.jl/issues/575
-    ("types_override.csv", (type=Int, types=Dict(:col1 => String)),
+    ("types_override.csv", (type=Int64, types=Dict(:col1 => String)),
         (3, 5),
         NamedTuple{(:col1,:col2,:col3,:col4,:col5), Tuple{String, Int64, Int64, Int64, Int64}},
         (col1 = ["A", "B", "C"], col2 = [1, 5, 9], col3 = [2, 6, 10], col4 = [3, 7, 11], col5 = [4, 8, 12])
